refactor: RuleSetのテストの配置場所を変更、namespaceの追加、JudgeResultTypeの変更
