+++ Assets/JJJ/Tests/Editor/Infrastructure/CpuHandStrategy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a40d6bd05b3c7f94094acd18a37c1f8e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
