+++ Assets/JJJ/Tests/Editor/Infrastructure/RuleSet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4122d88dfc8871c44a0cb6c958679aae
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
