+++ Assets/JJJ/Tests/Editor/Infrastructure/RuleSet/HardRuleSetTest.cs
@@ -2,7 +2,8 @@
 using JJJ.Infrastructure;
 using JJJ.Core.Entities;
 using NUnit.Framework;
-namespace JJJ.Tests.Infrastructure
+
+namespace JJJ.Tests.Infrastructure.RuleSet
 {
   /// <summary>
   /// HardRuleSetクラスのテスト
